requirements.txt
openai>=1.14.0
tqdm>=4.66.1
python-dotenv>=1.0.0
httpx>=0.27.0 